requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
pillow
tqdm
imbalanced-learn
matplotlib

